refactor(store): split segment label formatting into RecordingStore.SegmentLabels.cs
